Namespace Nsm mailers
