+++ app/mailers/nsm/feedback_messages/further_information_request_feedback.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+module Nsm
+  module FeedbackMessages
+    class FurtherInformationRequestFeedback < FeedbackBase
+      def template
+        '9ecdec30-83d7-468d-bec2-cf770a2c9828'
+      end
+
+      def contents
+        {
+          laa_case_reference: case_reference,
+          ufn: ufn,
+          main_defendant_name: defendant_name,
+          defendant_reference: defendant_reference_string,
+          claim_total: claim_total,
+          date_to_respond_by: 7.days.from_now.strftime('%d %B %Y'),
+          caseworker_information_requested: @comment,
+          date: DateTime.now.strftime('%d %B %Y'),
+          feedback_url: feedback_url
+        }
+      end
+    end
+  end
+end
