+++ spec/mailers/nsm/feedback_messages/rejected_feedback_spec.rb
@@ -2,17 +2,18 @@
 
 require 'rails_helper'
 
-RSpec.describe FeedbackMessages::ProviderRequestFeedback do
-  subject(:feedback) { described_class.new(claim) }
+RSpec.describe Nsm::FeedbackMessages::RejectedFeedback do
+  subject(:feedback) { described_class.new(claim, caseworker_decision_explanation) }
 
   let(:claim) { build(:claim) }
-  let(:feedback_template) { 'bfd28bd8-b9d8-4b18-8ce0-3fb763123573' }
+  let(:feedback_template) { '7e807120-b661-452c-95a6-1ae46f411cfe' }
   let(:recipient) { 'provider@example.com' }
   let(:laa_case_reference) { 'LAA-FHaMVK' }
   let(:ufn) { '123456/001' }
   let(:main_defendant_name) { 'Tracy Linklater' }
   let(:defendant_reference) { 'MAAT ID: AB12123' }
   let(:claim_total) { 0 }
+  let(:caseworker_decision_explanation) { 'Test Explanation' }
   let(:date) { DateTime.now.strftime('%d %B %Y') }
   let(:feedback_url) { 'tbc' }
 
@@ -30,6 +31,7 @@ RSpec.describe FeedbackMessages::ProviderRequestFeedback do
         main_defendant_name:,
         defendant_reference:,
         claim_total:,
+        caseworker_decision_explanation:,
         date:,
         feedback_url:
       )
